+++ wdevs-tax-switch.php
@@ -16,7 +16,7 @@
  * Plugin Name:          Tax Switch for WooCommerce
  * Plugin URI:           https://wijnberg.dev
  * Description:          Let customers toggle between inclusive and exclusive tax pricing in your WooCommerce store.
- * Version:              1.8.0
+ * Version:              1.8.1
  * Author:               Wijnberg Developments
  * Author URI:           https://wijnberg.dev/
  * License:              GPL-2.0+
@@ -40,7 +40,7 @@ if ( ! defined( 'WPINC' ) ) {
  * Start at version 1.0.0 and use SemVer - https://semver.org
  * Rename this for your plugin and update it as you release new versions.
  */
-define( 'WDEVS_TAX_SWITCH_VERSION', '1.8.0' );
+define( 'WDEVS_TAX_SWITCH_VERSION', '1.8.1' );
 
 /**
  * The code that runs during plugin activation.